Put the parse_argument outside of the class
