+++ src/eeg_research/simulators/decorators.py
@@ -8,7 +8,7 @@ from typing import Any, Callable, TypeVar, cast
 
 import bids
 
-from eeg_research.simulators.simulated_data import DummyDataset
+from eeg_research.simulators.simulate_data import DummyDataset
 
 FunctionType = TypeVar("FunctionType", bound=Callable[..., Any])
