+++ tests/simulators/test_cleaner_pipelines.py
@@ -13,26 +13,26 @@ import bids
 import pytest
 
 import eeg_research.simulators.cleaner_pipelines as script
-import eeg_research.simulators.simulated_data as simulated_data
+import eeg_research.simulators.simulate_data as simulate_data
 
 
 @pytest.fixture
-def dataset_structure() -> Generator[simulated_data.DummyDataset, None, None]:
+def dataset_structure() -> Generator[simulate_data.DummyDataset, None, None]:
     """Fixture to create a dataset object."""
     cwd = Path.cwd()
     output_dir = cwd.joinpath("data", "outputs")
     output_dir.mkdir(parents=True, exist_ok=True)
-    dataset_object = simulated_data.DummyDataset(root=output_dir, flush=False)
+    dataset_object = simulate_data.DummyDataset(root=output_dir, flush=False)
     yield dataset_object
 
 
 @pytest.fixture
-def light_dataset() -> Generator[simulated_data.DummyDataset, None, None]:
+def light_dataset() -> Generator[simulate_data.DummyDataset, None, None]:
     """Fixture to create a light dataset object."""
     cwd = Path.cwd()
     output_dir = cwd.joinpath("data", "outputs")
     output_dir.mkdir(parents=True, exist_ok=True)
-    dataset_object = simulated_data.DummyDataset(
+    dataset_object = simulate_data.DummyDataset(
         root=output_dir, task="test", flush=True
     )
     dataset_object.create_eeg_dataset(light=True, fmt="eeglab")
@@ -45,7 +45,7 @@ def heavy_dataset() -> Generator[script.CleanerPipelines, None, None]:
     cwd = Path.cwd()
     output_dir = cwd.joinpath("data", "outputs")
     output_dir.mkdir(parents=True, exist_ok=True)
-    dataset_object = simulated_data.DummyDataset(root=output_dir, flush=True)
+    dataset_object = simulate_data.DummyDataset(root=output_dir, flush=True)
     dataset_object.create_eeg_dataset(
         fmt="eeglab",
         n_channels=16,
@@ -68,7 +68,7 @@ def heavy_dataset() -> Generator[script.CleanerPipelines, None, None]:
     yield cleaner
 
 
-def test_append_message_to_txt_file(light_dataset: simulated_data.DummyDataset) -> None:
+def test_append_message_to_txt_file(light_dataset: simulate_data.DummyDataset) -> None:
     """Test that the function appends a message to a txt file."""
     bids_path = light_dataset.bids_path
     bids_layout = bids.layout.BIDSLayout(bids_path)
@@ -81,7 +81,7 @@ def test_append_message_to_txt_file(light_dataset: simulated_data.DummyDataset) 
         assert f.read() == message + "\n"
 
 
-def test_make_derivatives_path(light_dataset: simulated_data.DummyDataset) -> None:
+def test_make_derivatives_path(light_dataset: simulate_data.DummyDataset) -> None:
     """Test that the function creates the derivatives path."""
     bids_path = light_dataset.bids_path
     bids_layout = bids.layout.BIDSLayout(bids_path)
@@ -98,7 +98,7 @@ def test_make_derivatives_path(light_dataset: simulated_data.DummyDataset) -> No
     assert str(cleaner.derivatives_path) == str(expected_path)
 
 
-def test_make_process_path(light_dataset: simulated_data.DummyDataset) -> None:
+def test_make_process_path(light_dataset: simulate_data.DummyDataset) -> None:
     """Test that the function creates the process path."""
     bids_path = light_dataset.bids_path
     bids_layout = bids.layout.BIDSLayout(bids_path)
@@ -128,7 +128,7 @@ def test_make_process_path(light_dataset: simulated_data.DummyDataset) -> None:
         assert str(cleaner.process_path) == str(expected_path)
 
 
-def test_make_subject_session_path(light_dataset: simulated_data.DummyDataset) -> None:
+def test_make_subject_session_path(light_dataset: simulate_data.DummyDataset) -> None:
     """Test that the function creates the subject session path."""
     bids_path = light_dataset.bids_path
     bids_layout = bids.layout.BIDSLayout(bids_path)
@@ -158,7 +158,7 @@ def test_make_subject_session_path(light_dataset: simulated_data.DummyDataset) -
         assert str(cleaner.subject_session_path) == str(expected_path)
 
 
-def test_make_modality_path(light_dataset: simulated_data.DummyDataset) -> None:
+def test_make_modality_path(light_dataset: simulate_data.DummyDataset) -> None:
     """Test that the function creates the modality path."""
     bids_path = light_dataset.bids_path
     bids_layout = bids.layout.BIDSLayout(bids_path)
@@ -189,7 +189,7 @@ def test_make_modality_path(light_dataset: simulated_data.DummyDataset) -> None:
         assert str(cleaner.modality_path) == str(expected_path)
 
 
-def test_task_is_test(light_dataset: simulated_data.DummyDataset) -> None:
+def test_task_is_test(light_dataset: simulate_data.DummyDataset) -> None:
     """Test that the function returns True when the task is 'test'."""
     bids_path = light_dataset.bids_path
     bids_layout = bids.layout.BIDSLayout(bids_path)
@@ -203,7 +203,7 @@ def test_task_is_test(light_dataset: simulated_data.DummyDataset) -> None:
 
 
 def test_sidecar_copied_at_correct_location(
-    light_dataset: simulated_data.DummyDataset,
+    light_dataset: simulate_data.DummyDataset,
 ) -> None:
     """Test that the sidecar file is copied at the correct location."""
     bids_path = light_dataset.bids_path
@@ -229,13 +229,13 @@ def test_sidecar_copied_at_correct_location(
         assert expected_filename.exists()
 
 
-def test_save_raw_method(light_dataset: simulated_data.DummyDataset) -> None:
+def test_save_raw_method(light_dataset: simulate_data.DummyDataset) -> None:
     """Test that the function saves the raw data at the correct location."""
     bids_path = light_dataset.bids_path
     bids_layout = bids.layout.BIDSLayout(bids_path)
     bids_files = bids_layout.get(extension=".set")
     cleaner = script.CleanerPipelines(bids_files[0])
-    cleaner.raw = simulated_data.simulate_light_eeg_data()
+    cleaner.raw = simulate_data.simulate_light_eeg_data()
     cleaner.process_history = list()
     procedures = ["GRAD", "ASR", "PYPREP"]
     for procedure in procedures:
@@ -252,13 +252,13 @@ def test_save_raw_method(light_dataset: simulated_data.DummyDataset) -> None:
         assert os.path.isfile(expected_filename)
 
 
-def test_decorator_pipe(light_dataset: simulated_data.DummyDataset) -> None:
+def test_decorator_pipe(light_dataset: simulate_data.DummyDataset) -> None:
     """Test that the decorator pipe saves the raw data at the correct location."""
     bids_path = light_dataset.bids_path
     bids_layout = bids.layout.BIDSLayout(bids_path)
     bids_files = bids_layout.get(extension=".set")
     cleaner = script.CleanerPipelines(bids_files[0])
-    cleaner.raw = simulated_data.simulate_light_eeg_data()
+    cleaner.raw = simulate_data.simulate_light_eeg_data()
     procedures = "TEST_PIPE"
     cleaner.function_testing_decorator()
